+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Changed the Registry model to MonitoringLocation to better reflect that a registry (the app name) contains monitoring location instances.
 -   Set an order for lookups so that pick lists are in alphabetically order
 -   Using django-smart-selects to have chaining selects for state and county pick lists.
+-   Removed initialization of lookups from migrations and instead added a command to do this. This command will be run following migrations and will update any information that exists. This is also where the agency groups are created.
 
 ### Fixed
 -   Using whitenoise to serve out the staticfiles when using gunicorn to run the server. Added collectstaticfiles to the Dockerfile.